shell: allow specification of more complex shell commands
This allows users to specify any arbitrary program or command from the
command line. The only rule is that the raw ssh command is fed as the
last argument of the specified command.

For a handful of cases, like sh, bash, fish, zsh, pwsh, and powershell,
we still detect the right arguments that should be used.